cleanup migrate command
